Move the subprocess and vt100 text view into a new view, TerminalView.  This is in preparation for copying multiple-terminal support from trunk.
This also adds a dummy UIPageControl which may act as a terminal switcher.  The
status bar is used in the trunk, but it uses private APIs so this will work for
now (though we can revisit that if everyone likes it better).  This also adds an
info button which could be used to open up the preferences pane that is
currently triggered by the menu in trunk.